+++ src/components/writepage/mobile/WritePageMobilePosition.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, MouseEvent, useCallback } from 'react';
+import { ChangeEvent, MouseEvent } from 'react';
 import { mobilePositionList } from 'utils/positions';
 import {
   WritePageMobileBodyLeftBox,
@@ -9,32 +9,15 @@ import styled from '@emotion/styled';
 
 interface Props {
   positions: any;
-  setWriteFormValue: (value: any) => void;
+  onCalculateEvent: (e: MouseEvent<HTMLButtonElement>) => void;
   onFormValueChangeEvent: (e: ChangeEvent<HTMLInputElement>) => void;
 }
 
 const WritePageMobilePosition = ({
   positions,
-  setWriteFormValue,
+  onCalculateEvent,
   onFormValueChangeEvent,
 }: Props) => {
-  const handleCalculate = useCallback(
-    (e: MouseEvent<HTMLButtonElement>) => {
-      const { id, name, value } = e.currentTarget;
-      const numberValue = Number(value);
-      const updatedValue =
-        id === 'plus' ? numberValue + 1 : Math.max(0, numberValue - 1);
-      setWriteFormValue((prev: any) => ({
-        ...prev,
-        positions: {
-          ...prev.positions,
-          [name]: updatedValue,
-        },
-      }));
-    },
-    [setWriteFormValue],
-  );
-
   return (
     <WritePageMobilePositionContainer>
       <WritePageMobileBodyLeftBox>
@@ -48,7 +31,7 @@ const WritePageMobilePosition = ({
             position={position.name}
             value={positions[position.type]}
             onChangeEvent={onFormValueChangeEvent}
-            onClickEvent={handleCalculate}
+            onClickEvent={onCalculateEvent}
           />
         ))}
       </WritePageMobileBodyRightBox>
