+++ src/hooks/useSelectStack.ts
@@ -1,0 +1,77 @@
+import { ChangeEvent, Dispatch, useState } from 'react';
+import { WriteType } from 'types/write/writeType';
+
+interface Props {
+  plannerStack: string[];
+  designerStack: string[];
+  developerStack: string[];
+  setWriteFormValue: Dispatch<React.SetStateAction<WriteType.WriteFormType>>;
+}
+
+const useSelectStack = ({
+  plannerStack,
+  designerStack,
+  developerStack,
+  setWriteFormValue,
+}: Props) => {
+  const [selectedPlanner, setSelectedPlanner] = useState<string>('Figma');
+  const [selectedDesigner, setSelectedDesigner] = useState<string>('Figma');
+  const [selectedDeveloper, setSelectedDeveloper] =
+    useState<string>('JavaScript');
+
+  const handleSectionChange = (e: ChangeEvent<HTMLSelectElement>) => {
+    const { name, value } = e.target;
+    switch (name) {
+      case 'plannerStack':
+        setSelectedPlanner(value);
+        break;
+      case 'designerStack':
+        setSelectedDesigner(value);
+        break;
+      case 'developerStack':
+        setSelectedDeveloper(value);
+        break;
+      default:
+        break;
+    }
+  };
+
+  const handleAddStackButtonClick = (e: any) => {
+    const { name } = e.target;
+    switch (name) {
+      case 'plannerStack':
+        if (plannerStack.includes(selectedPlanner)) return;
+        setWriteFormValue((prev: any) => ({
+          ...prev,
+          plannerStack: [...prev.plannerStack, selectedPlanner],
+        }));
+        break;
+      case 'designerStack':
+        if (designerStack.includes(selectedDesigner)) return;
+        setWriteFormValue((prev: any) => ({
+          ...prev,
+          designerStack: [...prev.designerStack, selectedDesigner],
+        }));
+        break;
+      case 'developerStack':
+        if (developerStack.includes(selectedDeveloper)) return;
+        setWriteFormValue((prev: any) => ({
+          ...prev,
+          developerStack: [...prev.developerStack, selectedDeveloper],
+        }));
+        break;
+      default:
+        break;
+    }
+  };
+
+  return {
+    selectedPlanner,
+    selectedDesigner,
+    selectedDeveloper,
+    handleSectionChange,
+    handleAddStackButtonClick,
+  };
+};
+
+export default useSelectStack;
